+++ imageloader/src/main/java/ru/surfstudio/android/imageloader/ImageLoader.kt
@@ -133,8 +133,16 @@ class ImageLoader(private val context: Context) : ImageLoaderInterface {
      *
      * @param lambda лямбда, возвращающая загруженный [Drawable] и [ImageSource], указывающий откуда он был загружен
      */
-    override fun listener(lambda: ((drawable: Drawable, imageSource: ImageSource?) -> (Unit))) =
+    override fun listenerWithSource(lambda: ((drawable: Drawable, imageSource: ImageSource?) -> (Unit))) =
             apply { this.onImageLoadedLambda = lambda }
+
+    /**
+     * Установка лямбды для отслеживания загрузки изображения
+     *
+     * @param lambda лямбда, возвращающая загруженный [Drawable]
+     */
+    override fun listener(lambda: ((drawable: Drawable) -> (Unit))) =
+            apply { this.onImageLoadedLambda = { drawable, _ -> lambda(drawable) } }
 
     /**
      * Установка лямбды для отслеживания ошибки при загрузке изображения
@@ -143,7 +151,6 @@ class ImageLoader(private val context: Context) : ImageLoaderInterface {
      */
     override fun errorListener(lambda: ((throwable: Throwable) -> (Unit))) =
             apply { this.onImageLoadErrorLambda = lambda }
-
 
     /**
      * Указание политики кэширования.